refactor: move API-spec host helpers from strix.core.api_spec into interface.utils
